+++ R/simulate.R
@@ -285,7 +285,7 @@ simulate_vectorized = function(fit, ..., .type = "predict", .rate = FALSE, .dpar
   operation = switch(.type, fitted = "epred", loglik = "log_lik", predict = "rng")
   has_arma_terms = any(predictors$dpar %in% c("ar", "ma"))
   aux_operations = c(operation, if (has_arma_terms && .arma) "garma")
-  if (.type == "fitted" && .rate && (is.null(.dpar) || .dpar %in% c("epred", "mu")))
+  if (.type == "fitted" && (.rate || (!is.null(.dpar) && .dpar != "epred")))
     aux_operations = setdiff(aux_operations, "epred")
   aux_columns = get_family_aux_columns(fit$family, model_tables$segments, aux_operations)
   data_pars = c(data_columns$par_x, data_columns$series, stats::na.omit(unname(aux_columns)))
@@ -383,8 +383,11 @@ simulate_vectorized = function(fit, ..., .type = "predict", .rate = FALSE, .dpar
     return(dpar_values$mu_)
   }
 
-  if (.type == "fitted")
+  if (.type == "fitted") {
+    if (.dpar == "mu_")
+      return(dpar_values$mu_)
     return(fit$family$r$epred(dpars, response_data, rate = .rate))
+  }
   if (.type == "loglik")
     return(fit$family$r$log_lik(dpar_values$.ydata, dpars, response_data))
   if (.type == "predict") {